+++ auto_route/example/lib/web_demo/router/web_router.dart
@@ -3,7 +3,7 @@ import 'package:example/web_demo/router/web_router.gr.dart';
 import 'package:example/web_demo/web_main.dart';
 import 'package:flutter/material.dart';
 
-@AutoRouterConfig(generateForDir: ['lib/web_demo'],deferredLoading: false)
+@AutoRouterConfig(generateForDir: ['lib/web_demo'], deferredLoading: false)
 class WebAppRouter extends $WebAppRouter implements AutoRouteGuard {
   AuthService authService;
 
@@ -11,47 +11,43 @@ class WebAppRouter extends $WebAppRouter implements AutoRouteGuard {
 
   @override
   void onNavigation(NavigationResolver resolver, StackRouter router) {
-      if(authService.isAuthenticated || resolver.route.name == WebLoginRoute.name) {
-        resolver.next();
-      }else{
-        push(WebLoginRoute(resolver: resolver));
-      }
+    if (authService.isAuthenticated || resolver.route.name == WebLoginRoute.name) {
+      resolver.next();
+    } else {
+      push(WebLoginRoute(resolver: resolver));
+    }
   }
 
- @override
-   List<AutoRoute> get routes => [
-    AutoRoute(
-      page: MainWebRoute.page,
-      path: '/',
-      // guards: [AuthGuard(authService)],
-    ),
-    AutoRoute(
-      path: '/login',
-      page: WebLoginRoute.page,
-      keepHistory: false,
-    ),
-    AutoRoute(
-      path: '/user/:userID',
-      page: UserRoute.page,
-      children: [
-        AutoRoute(path: '', page: UserProfileRoute.page),
+  @override
+  List<AutoRoute> get routes => [
         AutoRoute(
-          path: 'posts',
-          page: UserPostsRoute.page,
+          page: MainWebRoute.page,
+          path: '/',
+          // guards: [AuthGuard(authService)],
+        ),
+        AutoRoute(
+          path: '/login',
+          page: WebLoginRoute.page,
+          keepHistory: false,
+        ),
+        AutoRoute(
+          path: '/user/:userID',
+          page: UserRoute.page,
           children: [
-            RedirectRoute(path: '', redirectTo: 'all'),
-            AutoRoute(path: 'all', page: UserAllPostsRoute.page),
-            AutoRoute(path: 'favorite', page: UserFavoritePostsRoute.page),
+            AutoRoute(path: '', page: UserProfileRoute.page),
+            AutoRoute(
+              path: 'posts',
+              page: UserPostsRoute.page,
+              children: [
+                RedirectRoute(path: '', redirectTo: 'all'),
+                AutoRoute(path: 'all', page: UserAllPostsRoute.page),
+                AutoRoute(path: 'favorite', page: UserFavoritePostsRoute.page),
+              ],
+            ),
           ],
         ),
-      ],
-    ),
-    AutoRoute(path: '*', page: NotFoundRoute.page),
-  ];
-
-
-
-
+        AutoRoute(path: '*', page: NotFoundRoute.page),
+      ];
 }
 
 @RoutePage()
@@ -86,8 +82,16 @@ class _MainWebPageState extends State<MainWebPage> {
             Padding(
               padding: const EdgeInsets.symmetric(vertical: 16),
               child: ElevatedButton(
-                onPressed:
-                    widget.navigate ?? () => context.navigateTo(UserRoute(id: 2, query: const ['value1', 'value2'])),
+                onPressed: widget.navigate ??
+                    () {
+                      context.pushRoute(
+                        UserRoute(
+                          id: 2,
+                          query: const ['value1', 'value2'],
+                        ),
+                      );
+
+                    },
                 child: Text('Navigate to user/2'),
               ),
             ),
@@ -159,7 +163,7 @@ class UserProfilePage extends StatelessWidget {
               color: Colors.red,
               onPressed: navigate ??
                   () {
-                    context.pushRoute( UserPostsRoute());
+                    context.pushRoute(UserPostsRoute());
                   },
               child: Text('Posts'),
             ),
@@ -181,6 +185,7 @@ class UserProfilePage extends StatelessWidget {
 @RoutePage()
 class UserPostsPage extends StatefulWidget {
   final int id;
+
   const UserPostsPage({@PathParam.inherit('userID') required this.id});
 
   @override