+++ auto_route/test/matcher_test.dart
@@ -1,6 +1,6 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:auto_route/src/matcher/route_matcher.dart';
-import 'package:flutter/widgets.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
@@ -23,11 +23,11 @@ void main() {
     );
 
     test('Creating root RouteCollection with a root route not starting with "/" should throw', () {
-      expect(()=> RouteCollection.from([TestRoute('A', path: 'a')], root: true), throwsFlutterError);
+      expect(() => RouteCollection.from([TestRoute('A', path: 'a')], root: true), throwsFlutterError);
     });
 
     test('Creating sub RouteCollection with a sub route starting with "/" should throw', () {
-      expect(()=> RouteCollection.from([TestRoute('A', path: '/a')], root: false), throwsFlutterError);
+      expect(() => RouteCollection.from([TestRoute('A', path: '/a')], root: false), throwsFlutterError);
     });
 
     test('Calling [routes] should return a list of all route configs', () {
@@ -66,25 +66,23 @@ void main() {
     });
     test('call RouteMatch.fromRedirect should return true', () {
       expect(
-          const RouteMatch(
-            name: 'A',
-            segments: ['a'],
-            path: 'a',
+          RouteMatch(
+            config: routeA,
+            segments: const ['a'],
             redirectedFrom: '/',
             stringMatch: 'a',
-            key: ValueKey('a'),
+            key: const ValueKey('a'),
           ).fromRedirect,
           isTrue);
     });
 
     test('call RouteMatch.hasEmptyPath should return true', () {
       expect(
-          const RouteMatch(
-            name: 'A',
-            segments: [''],
-            path: '',
+          RouteMatch(
+            config: TestRoute('A',path: ''),
+            segments: const [''],
             stringMatch: '',
-            key: ValueKey(''),
+            key: const ValueKey(''),
           ).hasEmptyPath,
           isTrue);
     });
@@ -113,12 +111,11 @@ void main() {
 
     test('Should return one match [B]', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'B',
-          key: ValueKey('B'),
-          path: '/b',
+        RouteMatch(
+          config: routeB,
+          key: const ValueKey('B'),
           stringMatch: '/b',
-          segments: ['/', 'b'],
+          segments: const ['/', 'b'],
         )
       ];
       expect(match('/b'), expectedMatches);
@@ -130,19 +127,17 @@ void main() {
 
     test('Should return one match with one nested match [C/C1]', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'C',
-          path: '/c',
+        RouteMatch(
+          config: routeC,
           stringMatch: '/c',
-          key: ValueKey('C'),
-          segments: ['/', 'c'],
+          key: const ValueKey('C'),
+          segments: const ['/', 'c'],
           children: [
             RouteMatch(
-              name: 'C1',
-              path: 'c1',
+              config: subRouteC1,
               stringMatch: 'c1',
-              key: ValueKey('C1'),
-              segments: ['c1'],
+              key: const ValueKey('C1'),
+              segments: const ['c1'],
             )
           ],
         )
@@ -182,19 +177,17 @@ void main() {
 
     test('Should return two matches [A,B]', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'A',
-          path: '/',
+        RouteMatch(
+          config: routeA,
           stringMatch: '/',
-          key: ValueKey('A'),
-          segments: ['/'],
+          key: const ValueKey('A'),
+          segments: const ['/'],
         ),
-        const RouteMatch(
-          name: 'B',
-          path: '/b',
+        RouteMatch(
+          config: routeB,
           stringMatch: '/b',
-          key: ValueKey('B'),
-          segments: ['/', 'b'],
+          key: const ValueKey('B'),
+          segments: const ['/', 'b'],
         )
       ];
       expect(match('/b', includePrefixMatches: true), expectedMatches);
@@ -202,26 +195,23 @@ void main() {
 
     test('Should return two prefix matches with one nested match [A, C/C1]', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'A',
-          path: '/',
+        RouteMatch(
+          config: routeA,
           stringMatch: '/',
-          key: ValueKey('A'),
-          segments: ['/'],
+          key: const ValueKey('A'),
+          segments: const ['/'],
         ),
-        const RouteMatch(
-          name: 'C',
-          path: '/c',
+        RouteMatch(
+          config: routeC,
           stringMatch: '/c',
-          key: ValueKey('C'),
-          segments: ['/', 'c'],
+          key: const ValueKey('C'),
+          segments: const ['/', 'c'],
           children: [
             RouteMatch(
-              name: 'C1',
-              path: 'c1',
+              config: subRouteC1,
               stringMatch: 'c1',
-              key: ValueKey('C1'),
-              segments: ['c1'],
+              key: const ValueKey('C1'),
+              segments: const ['c1'],
             )
           ],
         )
@@ -231,26 +221,23 @@ void main() {
 
     test('Should return two prefix matches with one nested match [A, D/D0]', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'A',
-          path: '/',
+        RouteMatch(
+          config: routeA,
           stringMatch: '/',
-          key: ValueKey('A'),
-          segments: ['/'],
+          key: const ValueKey('A'),
+          segments: const ['/'],
         ),
-        const RouteMatch(
-          name: 'D',
-          path: '/d',
+        RouteMatch(
+          config: routeD,
           stringMatch: '/d',
-          key: ValueKey('D'),
-          segments: ['/', 'd'],
+          key: const ValueKey('D'),
+          segments: const ['/', 'd'],
           children: [
             RouteMatch(
-              name: 'D0',
-              path: '',
+              config: subRouteD0,
               stringMatch: '',
-              key: ValueKey('D0'),
-              segments: [],
+              key: const ValueKey('D0'),
+              segments: const [],
             ),
           ],
         )
@@ -260,33 +247,29 @@ void main() {
 
     test('Should return two matches with two nested matches including empty path [A, D/D0/D1]', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'A',
-          path: '/',
+        RouteMatch(
+          config: routeA,
           stringMatch: '/',
-          key: ValueKey('A'),
-          segments: ['/'],
+          key: const ValueKey('A'),
+          segments: const ['/'],
         ),
-        const RouteMatch(
-          name: 'D',
-          path: '/d',
+        RouteMatch(
+          config: routeD,
           stringMatch: '/d',
-          key: ValueKey('D'),
-          segments: ['/', 'd'],
+          key: const ValueKey('D'),
+          segments: const ['/', 'd'],
           children: [
             RouteMatch(
-              path: '',
+              config: subRouteD0,
               stringMatch: '',
-              key: ValueKey('D0'),
-              name: 'D0',
-              segments: [],
+              key: const ValueKey('D0'),
+              segments: const [],
             ),
             RouteMatch(
-              name: 'D1',
+              config: subRouteD1,
               stringMatch: 'd1',
-              key: ValueKey('D1'),
-              path: 'd1',
-              segments: ['d1'],
+              key: const ValueKey('D1'),
+              segments: const ['d1'],
             )
           ],
         )
@@ -300,26 +283,23 @@ void main() {
 
     test('Should return two prefix matches and one full match [A, B, B/B1]', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'A',
-          key: ValueKey('A'),
+        RouteMatch(
+          config: routeA,
+          key: const ValueKey('A'),
           stringMatch: '/',
-          path: '/',
-          segments: ['/'],
+          segments: const ['/'],
         ),
-        const RouteMatch(
-          name: 'B',
-          key: ValueKey('B'),
+        RouteMatch(
+          config: routeB,
+          key: const ValueKey('B'),
           stringMatch: '/b',
-          path: '/b',
-          segments: ['/', 'b'],
+          segments: const ['/', 'b'],
         ),
-        const RouteMatch(
-          name: 'B1',
-          key: ValueKey('B1'),
+        RouteMatch(
+          config: routeB1,
+          key: const ValueKey('B1'),
           stringMatch: '/b/b1',
-          path: '/b/b1',
-          segments: ['/', 'b', 'b1'],
+          segments: const ['/', 'b', 'b1'],
         )
       ];
       expect(match('/b/b1', includePrefixMatches: true), expectedMatches);
@@ -346,12 +326,11 @@ void main() {
 
     test('Should match wildcard [WC]', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'WC',
-          key: ValueKey('WC'),
+        RouteMatch(
+          config: wcRoute,
+          key: const ValueKey('WC'),
           stringMatch: '/x/y',
-          path: '*',
-          segments: ['/', 'x', 'y'],
+          segments: const ['/', 'x', 'y'],
         )
       ];
       expect(match('/x/y', includePrefixMatches: true), expectedMatches);
@@ -359,12 +338,11 @@ void main() {
 
     test('Incomplete match should return [WC]', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'WC',
-          key: ValueKey('WC'),
+        RouteMatch(
+          config: wcRoute,
+          key: const ValueKey('WC'),
           stringMatch: '/c/c1/x',
-          path: '*',
-          segments: ['/', 'c', 'c1', 'x'],
+          segments: const ['/', 'c', 'c1', 'x'],
         )
       ];
       expect(match('/c/c1/x', includePrefixMatches: true), expectedMatches);
@@ -372,12 +350,11 @@ void main() {
 
     test('Should match prefixed wildcard route [PWC]', () {
       final expectedMatches = [
-        const RouteMatch(
-          path: '/d/*',
-          key: ValueKey('PWC'),
+        RouteMatch(
+          config: prefixedWcRoute,
+          key: const ValueKey('PWC'),
           stringMatch: '/d/x/y',
-          name: 'PWC',
-          segments: ['/', 'd', 'x', 'y'],
+          segments: const ['/', 'd', 'x', 'y'],
         )
       ];
       expect(match('/d/x/y'), expectedMatches);
@@ -406,12 +383,11 @@ void main() {
 
     test('Should match route [A]', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'A',
-          key: ValueKey('A'),
+        RouteMatch(
+          config: routeA,
+          key: const ValueKey('A'),
           stringMatch: '/a',
-          path: '/a',
-          segments: ['/', 'a'],
+          segments: const ['/', 'a'],
           redirectedFrom: '/',
         )
       ];
@@ -420,19 +396,17 @@ void main() {
 
     test('Should match route [C/C1]', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'C',
-          key: ValueKey('C'),
+        RouteMatch(
+          config: routeC,
+          key: const ValueKey('C'),
           stringMatch: '/c',
-          path: '/c',
-          segments: ['/', 'c'],
+          segments: const ['/', 'c'],
           children: [
             RouteMatch(
-              name: 'C1',
-              key: ValueKey('C1'),
+              config: subRouteC1,
+              key: const ValueKey('C1'),
               stringMatch: 'c1',
-              path: 'c1',
-              segments: ['c1'],
+              segments: const ['c1'],
               redirectedFrom: '',
             )
           ],
@@ -455,19 +429,17 @@ void main() {
 
     test('Should match route [A/A1] subRedirect to empty path', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'A',
-          key: ValueKey('A'),
+        RouteMatch(
+          config: routeA,
+          key: const ValueKey('A'),
           stringMatch: '/a',
-          path: '/a',
-          segments: ['/', 'a'],
+          segments: const ['/', 'a'],
           children: [
             RouteMatch(
-              name: 'A1',
-              key: ValueKey('A1'),
+              config: TestRoute('A1', path: ''),
+              key: const ValueKey('A1'),
               stringMatch: '',
-              path: '',
-              segments: [],
+              segments: const [],
               redirectedFrom: 'r',
             )
           ],
@@ -496,13 +468,12 @@ void main() {
 
     test('Should match route [A] and extract path param {id:1}', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'A',
-          key: ValueKey('A'),
+        RouteMatch(
+          config: routeA,
+          key: const ValueKey('A'),
           stringMatch: '/a/1',
-          path: '/a/:id',
-          segments: ['/', 'a', '1'],
-          pathParams: Parameters({'id': '1'}),
+          segments: const ['/', 'a', '1'],
+          pathParams: const Parameters({'id': '1'}),
         )
       ];
       expect(match('/a/1'), expectedMatches);
@@ -510,13 +481,12 @@ void main() {
 
     test('Should match route [B] and extract path params {id:1, type:none}', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'B',
-          key: ValueKey('B'),
+        RouteMatch(
+          config: routeB,
+          key: const ValueKey('B'),
           stringMatch: '/b/1/n/none',
-          path: '/b/:id/n/:type',
-          segments: ['/', 'b', '1', 'n', 'none'],
-          pathParams: Parameters({
+          segments: const ['/', 'b', '1', 'n', 'none'],
+          pathParams: const Parameters({
             'id': '1',
             'type': 'none',
           }),
@@ -527,20 +497,18 @@ void main() {
 
     test('Should match route [C]', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'C',
-          key: ValueKey('C'),
+        RouteMatch(
+          config: routeC,
+          key: const ValueKey('C'),
           stringMatch: '/c',
-          path: '/c',
-          segments: ['/', 'c'],
+          segments: const ['/', 'c'],
           children: [
             RouteMatch(
-              path: ':id',
-              name: 'C1',
-              key: ValueKey('C1'),
+             config: subRouteC1,
+              key: const ValueKey('C1'),
               stringMatch: '1',
-              segments: ['1'],
-              pathParams: Parameters({'id': '1'}),
+              segments: const ['1'],
+              pathParams: const Parameters({'id': '1'}),
             )
           ],
         )
@@ -568,13 +536,12 @@ void main() {
 
     test('Should match route [A] and extract query param {foo:bar}', () {
       final expectedMatches = [
-        const RouteMatch(
-          key: ValueKey('A'),
+        RouteMatch(
+          key: const ValueKey('A'),
           stringMatch: '/a',
-          name: 'A',
-          path: '/a',
-          segments: ['/', 'a'],
-          queryParams: Parameters({'foo': 'bar'}),
+          config: routeA,
+          segments: const ['/', 'a'],
+          queryParams: const Parameters({'foo': 'bar'}),
         )
       ];
       expect(match('/a?foo=bar'), expectedMatches);
@@ -582,21 +549,19 @@ void main() {
 
     test('Should match routes [B,B1] and extract query params {foo:bar, bar:baz} for both', () {
       final expectedMatches = [
-        const RouteMatch(
-          key: ValueKey('B'),
+        RouteMatch(
+          key: const ValueKey('B'),
           stringMatch: '/b',
-          name: 'B',
-          path: '/b',
-          segments: ['/', 'b'],
-          queryParams: Parameters({'foo': 'bar', 'bar': 'baz'}),
+          config: routeB,
+          segments: const ['/', 'b'],
+          queryParams: const Parameters({'foo': 'bar', 'bar': 'baz'}),
         ),
-        const RouteMatch(
-          key: ValueKey('B1'),
+        RouteMatch(
+          key: const ValueKey('B1'),
           stringMatch: '/b/b1',
-          name: 'B1',
-          path: '/b/b1',
-          segments: ['/', 'b', 'b1'],
-          queryParams: Parameters({'foo': 'bar', 'bar': 'baz'}),
+          config: routeB1,
+          segments: const ['/', 'b', 'b1'],
+          queryParams: const Parameters({'foo': 'bar', 'bar': 'baz'}),
         )
       ];
       expect(match('/b/b1?foo=bar&bar=baz', includePrefixMatches: true), expectedMatches);
@@ -604,21 +569,19 @@ void main() {
 
     test('Should match route [C/C1] and extract query parameters {foo:bar} for parent and child', () {
       final expectedMatches = [
-        const RouteMatch(
-          name: 'C',
-          path: '/c',
+        RouteMatch(
+          config: routeC,
           stringMatch: '/c',
-          key: ValueKey('C'),
-          segments: ['/', 'c'],
-          queryParams: Parameters({'foo': 'bar'}),
+          key: const ValueKey('C'),
+          segments: const ['/', 'c'],
+          queryParams: const Parameters({'foo': 'bar'}),
           children: [
             RouteMatch(
-              key: ValueKey('C1'),
-              name: 'C1',
-              path: 'c1',
+              key: const ValueKey('C1'),
+              config: subRouteC1,
               stringMatch: 'c1',
-              segments: ['c1'],
-              queryParams: Parameters({'foo': 'bar'}),
+              segments: const ['c1'],
+              queryParams: const Parameters({'foo': 'bar'}),
             )
           ],
         )
@@ -628,13 +591,12 @@ void main() {
 
     test('Should match route [A] and extract query param {foo:[bar,baz]}', () {
       final expectedMatches = [
-        const RouteMatch(
-          key: ValueKey('A'),
+        RouteMatch(
+          key: const ValueKey('A'),
           stringMatch: '/a',
-          name: 'A',
-          path: '/a',
-          segments: ['/', 'a'],
-          queryParams: Parameters({
+          config: routeA,
+          segments: const ['/', 'a'],
+          queryParams: const Parameters({
             'foo': ['bar', 'baz']
           }),
         )