+++ auto_route/test/nested_tabs_router/router.dart
@@ -1,4 +1,5 @@
 import 'package:auto_route/auto_route.dart';
+import 'package:auto_route/src/route/auto_route_config.dart';
 
 import '../main_router.dart';