+++ auto_route/CHANGELOG.md
@@ -1,6 +1,8 @@
 # ChangeLog
 ## [6.0.6]
 - RootRouter can implement AutoRouteGuard and guard all Stack-routes
+- Add restorationId builder
+- Mark annotations with supported types
 ## [6.0.5]
 - Fix RouteMatch.fullPath is not joined properly
 - Add helper method to get route match PageRouteInfo().match(context)